Version1
This is the work i have completed until now for abercromby vending app. So far I have only done work within the css and index.html files within the www folder. This is a basic login page with no functionality as i still need to complete a tutorial to find out how to use php with jquery/ajax as phone gap does not allow php files.