newest version
This code has a main page with a nav bar that will allow the user to navigate around the app.
As well as this there are 6 objects users will be able to view or buy. So far they can see the details of each project. This isnt done very efficiently as it uses multiple php pages to display this information, however, it made for a quick and easy resolution so that I had on example of a possible feature that abrcromby may want to consider when they go on to create their real thing.