Added some functionality and html
The logout page should clear whats stored in local storage. I still need to figure out how to use whats stored in local storage to skip the log in page. I store the username as a main way of deciphering between users bc username has to be unique.
Started on an accounts page. Allow the user to select from drop down menu and upload into their database. I am not going to ask for any bank details etc as this would be time consuming and thomas isnt too fussed about this. this is mainly to show how the health points will work. I will also try and use push notifications to alert the user when their balance is low as an extra feature for the app.